Refactor QuotationService and PromotionHelper; enhance quotation creation logic with detailed pricing and promotion handling; remove unused DealerPointUtil; update application configuration for database connection.

+++ src/main/java/com/example/emob/entity/ElectricVehicle.java
@@ -3,6 +3,7 @@ package com.example.emob.entity;
 
 import com.example.emob.constant.VehicleType;
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.List;
@@ -23,8 +24,8 @@ public class ElectricVehicle {
 
   String brand;
   String model;
-  float importPrice;
-  float retailPrice;
+  BigDecimal importPrice;
+  BigDecimal retailPrice;
   Float batteryKwh;
   Integer rangeKm;
   Float chargeTimeHr;
