+++ src/main/java/com/example/emob/model/request/vehicle/ElectricVehiclePriceRequest.java
@@ -1,6 +1,7 @@
 /* EMOB-2025 */
 package com.example.emob.model.request.vehicle;
 
+import java.math.BigDecimal;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
@@ -10,6 +11,6 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ElectricVehiclePriceRequest {
-  float importPrice;
-  float retailPrice;
+  BigDecimal importPrice;
+  BigDecimal retailPrice;
 }
